Revert "tools/automake: bump to 1.17"
This reverts commit ac4040ab0648c12835781d4d2e97407ca2b462ce.

Bump to 1.17 broke SDK portability as it partially reverted commit
("automake: portability fixes") which made Perl PATH "/usr/bin/env perl"
as that is portable, as otherwise automake will set the absolute path to
the Perl binary and this will then fail when using SDK on a different
system as that PATH is not present.

So, since fixing this would require backport of upstream commit
("configure: make perl path with whitespace a warning, not error.") which
requires autoreconf to be done in order for configure to get regenerated
we cannot do it because at that time we do not have automake built.

So, for now revert the bump until upstream makes a new release.

Signed-off-by: Robert Marko <robimarko@gmail.com>